+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.9.4 (March 20, 2026)
 
 IMPROVEMENTS
-* Bump Consul image to `1.21.5` and `1.21.9-ent`
+* Bump Consul image to `1.21.5` and `1.21.12-ent`
 * Bump Consul ECS image to `0.9.4`
 * Bump Go version to `1.25.8`